Use validated device provider configs for public status
