+++ packages/device-syncd/src/config/provider-manifests.ts
@@ -15,58 +15,24 @@ import { configuredDeviceSyncProviderCredentialPolicies } from "../provider-cred
 import { configuredDeviceSyncProviderJobDefinitions } from "../provider-job-definitions.ts";
 
 import { createJunctionDeviceSyncProvider } from "../providers/junction.ts";
-import { readConfiguredJunctionDeviceSyncProviderConfig } from "../providers/junction-config.ts";
 import { createOuraDeviceSyncProvider } from "../providers/oura.ts";
 import { createStravaDeviceSyncProvider } from "../providers/strava.ts";
 import { createWhoopDeviceSyncProvider } from "../providers/whoop.ts";
-
 import {
-  optionalEnv,
-  parseCsvEnv,
-  parseIntegerEnv,
-  readOptionalCredentialPair,
-} from "./provider-config-helpers.ts";
+  readConfiguredJunctionDeviceSyncProviderConfig,
+  readConfiguredOuraDeviceSyncProviderConfig,
+  readConfiguredStravaDeviceSyncProviderConfig,
+  readConfiguredWhoopDeviceSyncProviderConfig,
+} from "./provider-configs.ts";
 import {
   configuredDeviceSyncProviderKeys,
   listConfiguredDeviceSyncProviderNames,
 } from "./provider-keys.ts";
 import {
   JUNCTION_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  OURA_API_BASE_URL_ENV_KEYS,
-  OURA_AUTH_BASE_URL_ENV_KEYS,
-  OURA_BACKFILL_DAYS_ENV_KEYS,
-  OURA_CLIENT_ID_ENV_KEYS,
-  OURA_CLIENT_SECRET_ENV_KEYS,
   OURA_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  OURA_RECONCILE_DAYS_ENV_KEYS,
-  OURA_RECONCILE_INTERVAL_MS_ENV_KEYS,
-  OURA_REQUEST_TIMEOUT_MS_ENV_KEYS,
-  OURA_SCOPES_ENV_KEYS,
-  OURA_WEBHOOK_TIMESTAMP_TOLERANCE_MS_ENV_KEYS,
-  OURA_WEBHOOK_VERIFICATION_TOKEN_ENV_KEYS,
-  STRAVA_API_BASE_URL_ENV_KEYS,
-  STRAVA_AUTH_BASE_URL_ENV_KEYS,
-  STRAVA_BACKFILL_DAYS_ENV_KEYS,
-  STRAVA_CLIENT_ID_ENV_KEYS,
-  STRAVA_CLIENT_SECRET_ENV_KEYS,
   STRAVA_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  STRAVA_RECONCILE_DAYS_ENV_KEYS,
-  STRAVA_RECONCILE_INTERVAL_MS_ENV_KEYS,
-  STRAVA_REQUEST_TIMEOUT_MS_ENV_KEYS,
-  STRAVA_SCOPES_ENV_KEYS,
-  STRAVA_WEBHOOK_SIGNING_SECRET_ENV_KEYS,
-  STRAVA_WEBHOOK_TIMESTAMP_TOLERANCE_MS_ENV_KEYS,
-  STRAVA_WEBHOOK_VERIFY_TOKEN_ENV_KEYS,
-  WHOOP_BACKFILL_DAYS_ENV_KEYS,
-  WHOOP_BASE_URL_ENV_KEYS,
-  WHOOP_CLIENT_ID_ENV_KEYS,
-  WHOOP_CLIENT_SECRET_ENV_KEYS,
   WHOOP_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  WHOOP_RECONCILE_DAYS_ENV_KEYS,
-  WHOOP_RECONCILE_INTERVAL_MS_ENV_KEYS,
-  WHOOP_REQUEST_TIMEOUT_MS_ENV_KEYS,
-  WHOOP_SCOPES_ENV_KEYS,
-  WHOOP_WEBHOOK_TIMESTAMP_TOLERANCE_MS_ENV_KEYS,
   type DeviceSyncProviderEnvSpec,
 } from "./provider-env.ts";
 
@@ -212,32 +178,7 @@ const OURA_DEVICE_SYNC_PROVIDER_MANIFEST = defineConfiguredDeviceSyncProviderMan
   descriptor: configuredDeviceSyncProviderDescriptors.oura,
   importer: ouraProviderAdapter,
   env: OURA_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  readConfig(env) {
-    const credentials = readOptionalCredentialPair(
-      env,
-      OURA_CLIENT_ID_ENV_KEYS,
-      OURA_CLIENT_SECRET_ENV_KEYS,
-      "Oura",
-    );
-
-    if (!credentials) {
-      return null;
-    }
-
-    return {
-      clientId: credentials.clientId,
-      clientSecret: credentials.clientSecret,
-      authBaseUrl: optionalEnv(env, OURA_AUTH_BASE_URL_ENV_KEYS),
-      apiBaseUrl: optionalEnv(env, OURA_API_BASE_URL_ENV_KEYS),
-      scopes: parseCsvEnv(env, OURA_SCOPES_ENV_KEYS),
-      backfillDays: parseIntegerEnv(env, OURA_BACKFILL_DAYS_ENV_KEYS),
-      reconcileDays: parseIntegerEnv(env, OURA_RECONCILE_DAYS_ENV_KEYS),
-      reconcileIntervalMs: parseIntegerEnv(env, OURA_RECONCILE_INTERVAL_MS_ENV_KEYS),
-      webhookTimestampToleranceMs: parseIntegerEnv(env, OURA_WEBHOOK_TIMESTAMP_TOLERANCE_MS_ENV_KEYS),
-      requestTimeoutMs: parseIntegerEnv(env, OURA_REQUEST_TIMEOUT_MS_ENV_KEYS),
-      webhookVerificationToken: optionalEnv(env, OURA_WEBHOOK_VERIFICATION_TOKEN_ENV_KEYS),
-    };
-  },
+  readConfig: readConfiguredOuraDeviceSyncProviderConfig,
   createProvider: createOuraDeviceSyncProvider,
   serializableFields: {
     apiBaseUrl: "string",
@@ -269,30 +210,7 @@ const WHOOP_DEVICE_SYNC_PROVIDER_MANIFEST = defineConfiguredDeviceSyncProviderMa
   descriptor: configuredDeviceSyncProviderDescriptors.whoop,
   importer: whoopProviderAdapter,
   env: WHOOP_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  readConfig(env) {
-    const credentials = readOptionalCredentialPair(
-      env,
-      WHOOP_CLIENT_ID_ENV_KEYS,
-      WHOOP_CLIENT_SECRET_ENV_KEYS,
-      "WHOOP",
-    );
-
-    if (!credentials) {
-      return null;
-    }
-
-    return {
-      clientId: credentials.clientId,
-      clientSecret: credentials.clientSecret,
-      baseUrl: optionalEnv(env, WHOOP_BASE_URL_ENV_KEYS),
-      scopes: parseCsvEnv(env, WHOOP_SCOPES_ENV_KEYS),
-      backfillDays: parseIntegerEnv(env, WHOOP_BACKFILL_DAYS_ENV_KEYS),
-      reconcileDays: parseIntegerEnv(env, WHOOP_RECONCILE_DAYS_ENV_KEYS),
-      reconcileIntervalMs: parseIntegerEnv(env, WHOOP_RECONCILE_INTERVAL_MS_ENV_KEYS),
-      webhookTimestampToleranceMs: parseIntegerEnv(env, WHOOP_WEBHOOK_TIMESTAMP_TOLERANCE_MS_ENV_KEYS),
-      requestTimeoutMs: parseIntegerEnv(env, WHOOP_REQUEST_TIMEOUT_MS_ENV_KEYS),
-    };
-  },
+  readConfig: readConfiguredWhoopDeviceSyncProviderConfig,
   createProvider: createWhoopDeviceSyncProvider,
   serializableFields: {
     backfillDays: "number",
@@ -319,33 +237,7 @@ const STRAVA_DEVICE_SYNC_PROVIDER_MANIFEST = defineConfiguredDeviceSyncProviderM
   descriptor: configuredDeviceSyncProviderDescriptors.strava,
   importer: stravaProviderAdapter,
   env: STRAVA_DEVICE_SYNC_PROVIDER_ENV_SPEC,
-  readConfig(env) {
-    const credentials = readOptionalCredentialPair(
-      env,
-      STRAVA_CLIENT_ID_ENV_KEYS,
-      STRAVA_CLIENT_SECRET_ENV_KEYS,
-      "Strava",
-    );
-
-    if (!credentials) {
-      return null;
-    }
-
-    return {
-      clientId: credentials.clientId,
-      clientSecret: credentials.clientSecret,
-      authBaseUrl: optionalEnv(env, STRAVA_AUTH_BASE_URL_ENV_KEYS),
-      apiBaseUrl: optionalEnv(env, STRAVA_API_BASE_URL_ENV_KEYS),
-      scopes: parseCsvEnv(env, STRAVA_SCOPES_ENV_KEYS),
-      backfillDays: parseIntegerEnv(env, STRAVA_BACKFILL_DAYS_ENV_KEYS),
-      reconcileDays: parseIntegerEnv(env, STRAVA_RECONCILE_DAYS_ENV_KEYS),
-      reconcileIntervalMs: parseIntegerEnv(env, STRAVA_RECONCILE_INTERVAL_MS_ENV_KEYS),
-      requestTimeoutMs: parseIntegerEnv(env, STRAVA_REQUEST_TIMEOUT_MS_ENV_KEYS),
-      webhookSigningSecret: optionalEnv(env, STRAVA_WEBHOOK_SIGNING_SECRET_ENV_KEYS),
-      webhookTimestampToleranceMs: parseIntegerEnv(env, STRAVA_WEBHOOK_TIMESTAMP_TOLERANCE_MS_ENV_KEYS),
-      webhookVerifyToken: optionalEnv(env, STRAVA_WEBHOOK_VERIFY_TOKEN_ENV_KEYS),
-    };
-  },
+  readConfig: readConfiguredStravaDeviceSyncProviderConfig,
   createProvider: createStravaDeviceSyncProvider,
   serializableFields: {
     apiBaseUrl: "string",
