+++ packages/device-syncd/src/provider-configs.ts
@@ -1,0 +1,18 @@
+export {
+  configuredDeviceSyncProviderKeys,
+  hasConfiguredDeviceSyncProviderConfigs,
+  listConfiguredDeviceSyncProviderNames,
+  readConfiguredDeviceSyncProviderConfigs,
+  readConfiguredJunctionDeviceSyncProviderConfig,
+  readConfiguredOuraDeviceSyncProviderConfig,
+  readConfiguredStravaDeviceSyncProviderConfig,
+  readConfiguredWhoopDeviceSyncProviderConfig,
+} from "./config/provider-configs.ts";
+
+export type {
+  ConfiguredDeviceSyncProviderConfigByKey,
+  ConfiguredDeviceSyncProviderConfigs,
+  ConfiguredDeviceSyncProviderKey,
+  ConfiguredDeviceSyncProviderPresence,
+  DeviceSyncEnvSource,
+} from "./config/provider-configs.ts";
